Move pathDB cleaner to pathdb package
It makes more sense to have it in the same package.
